Fixed coauthor and author functionality
