+++ App/models/__init__.py
@@ -1,2 +1,4 @@
 from .user import *
+from .author import *
+from .publication import *
 from .model import *
